Audio crash in 1.09 fix(?) and small cleanup
Originally committed to SVN as r17.

+++ core/audio_display.h
@@ -66,7 +66,7 @@ private:
 	__int64 PositionSample;
 	float scale;
 	int samples;
-	int Position;
+	__int64 Position;
 	int samplesPercent;
 	int oldCurPos;
 	bool hasFocus;
